refactor serde
