+++ src/stream/src/executor/lookup/tests.rs
@@ -19,9 +19,10 @@ use risingwave_common::array::stream_chunk::StreamChunkTestExt;
 use risingwave_common::array::StreamChunk;
 use risingwave_common::catalog::{ColumnDesc, ColumnId, Field, Schema, TableId};
 use risingwave_common::types::DataType;
-use risingwave_common::util::ordered::{deserialize_column_id, SENTINEL_CELL_ID};
+use risingwave_common::util::ordered::SENTINEL_CELL_ID;
 use risingwave_common::util::sort_util::{OrderPair, OrderType};
 use risingwave_common::util::value_encoding::deserialize_cell;
+use risingwave_storage::encoding::cell_based_encoding_util::deserialize_column_id;
 use risingwave_storage::memory::MemoryStateStore;
 use risingwave_storage::store::ReadOptions;
 use risingwave_storage::StateStore;
